+++ conftest.py
@@ -3,6 +3,7 @@ import configparser
 import re
 import subprocess
 import time
+import uuid
 
 import psutil
 import pytest
@@ -13,6 +14,10 @@ NGROK_CLI = [
 ]
 
 def pytest_addoption(parser):
+    parser.addoption('--addons-path')
+    parser.addoption('--db', help="DB to run the tests against", default=str(uuid.uuid4()))
+    parser.addoption("--no-delete", action="store_true", help="Don't delete repo after a failed run")
+
     parser.addoption(
         '--tunnel', action="store", type="choice", choices=['ngrok', 'localtunnel'], default='ngrok',
         help="Which tunneling method to use to expose the local Odoo server "
@@ -22,6 +27,9 @@ def pytest_addoption(parser):
              "queries per minute, free is 40, multi-repo batching tests will "
              "blow through the former); localtunnel has no rate-limiting but "
              "the servers are way less reliable")
+
+def pytest_report_header(config):
+    return 'Running against database ' + config.getoption('--db')
 
 @pytest.fixture(scope="session")
 def config(pytestconfig):
@@ -140,3 +148,29 @@ def tunnel(pytestconfig, port):
             p.wait(30)
     else:
         raise ValueError("Unsupported %s tunnel method" % tunnel)
+
+@pytest.fixture(scope='session')
+def dbcache(request, module):
+    """ Creates template DB once per run, then just duplicates it before
+    starting odoo and running the testcase
+    """
+    db = request.config.getoption('--db')
+    subprocess.run([
+        'odoo', '--no-http',
+        '--addons-path', request.config.getoption('--addons-path'),
+        '-d', db, '-i', module,
+        '--max-cron-threads', '0',
+        '--stop-after-init'
+    ], check=True)
+    yield db
+    subprocess.run(['dropdb', db])
+
+@pytest.fixture
+def db(request, dbcache):
+    rundb = str(uuid.uuid4())
+    subprocess.run(['createdb', '-T', dbcache, rundb], check=True)
+
+    yield rundb
+
+    if not request.config.getoption('--no-delete'):
+        subprocess.run(['dropdb', rundb], check=True)
