+++ runbot_merge/tests/conftest.py
@@ -1,5 +1,8 @@
+import uuid
+import pytest
+
 pytest_plugins = ["local"]
 
-def pytest_addoption(parser):
-    parser.addoption("--db", action="store", help="Odoo DB to run tests with")
-    parser.addoption('--addons-path', action='store', help="Odoo's addons path")
+@pytest.fixture(scope='session')
+def module():
+    return 'runbot_merge'
